--- frontend/src/views/UserProfile.vue
@@ -1,100 +1,0 @@
-<template>
-    <div class="user-profile">
-        <v-container grid-list>
-            <v-layout row>
-                <v-layout class="user-data" column align-center justify-center ma-4 v-if="showProfile">
-
-                    <v-avatar :tile="false" :size="192" color="grey lighten-4">
-                        <img src="@/assets/img/user.png" alt="avatar">
-                    </v-avatar>
-
-                    <h2 class="text-xs-center">{{ user.firstName + ' ' + user.lastName }}</h2>
-                    <h4 class="text-xs-center">{{ user.email }}</h4>
-                    <h4 class="text-xs-center">{{ user.phone }}</h4>
-                    <h4 class="text-xs-center">{{ user.address }}</h4>
-
-                    <div class="text-xs-center">
-                        <v-btn v-on:click="changeToEdit()" round small flat color="primary" dark>Edit Profile</v-btn>
-                    </div>
-                    <v-spacer></v-spacer>
-                </v-layout>
-                <component v-bind:is="activeComponent"
-                   v-on:showFriendshipsClicked="changeToFriendships($event)"
-                   v-on:searchUsersClicked="changeToSearch($event)"
-                   v-on:showFriendRequestsClicked="changeToFriendRequests($event)"
-            />
-            </v-layout>
-
-        </v-container>
-
-    </div>
-</template>
-
-<script>
-import EditUser from "@/components/EditUser.vue"
-import Friendships from "@/components/Friendships.vue"
-import SearchUsers from "@/components/SearchUsers.vue"
-import FriendRequests from "@/components/FriendRequests.vue"
-
-export default {
-    components: {
-        'edit-user' : EditUser,
-        'friendships': Friendships,
-        'search-users' : SearchUsers,
-        'friend-requests': FriendRequests
-    },
-    props: {
-        userAvatar: {
-            type: String,
-            default: require("@/assets/img/user.png")
-        }
-    },
-    data() {
-        return {
-            activeComponent: 'friendships',
-            showProfile: true,
-            user: {
-                firstName: '',
-                lastName: '',
-                email: '',
-                phone: '',
-                address: ''
-            }
-        };
-    },
-    methods: {
-        changeToEdit: function() {
-            this.activeComponent = 'edit-user';
-            this.showProfile = true;
-        },
-        changeToSearch: function() {
-            this.activeComponent = 'search-users';
-            this.showProfile = false;
-        },
-        changeToFriendRequests: function() {
-            this.activeComponent = 'friend-requests';
-            this.showProfile = true;
-        }
-    },
-    created() {
-
-                var yourConfig = {
-            headers: {
-                Authorization: "Bearer " + localStorage.getItem("token")
-            }
-        }
-
-        this.$axios.get('http://localhost:8080/api/currentUser/', yourConfig).then((response) => {
-            this.user = response.data;
-        }).catch(function(error) {
-                alert(error.response.data.message);
-            });
-    }
-}
-</script>
-
-<style>
-.user-data {
-    max-width: 800px;
-}
-</style>
